Added dummy AFSs; Changed Erdogan (Fuel Constraints) to check only if
the next AFS is within reach from any Node. The old way (checking if
next AFS or destination depot is in reach) could lead to prolems if the
destination depot is closer than the next AFS, but the route used the
AFS. The fuel level z could fall below zero if this happens. This
problem is not covered by the upper and lower bounds of z, because they
are checked after a node is visited. The fuel level before the AFS is
somerwhere below zero (which would lead to problems). After the AFS it
is on the maximum (which is in the bounds of z) and therefore will not
lead to any problems.

The model works, but in cause of 5 more nodes it takes about a minute to
solve.
The model contains now 20 Nodes and 5 Trucks.